Moved utility functions into admin.utils package.
