+++ admin/dump_json.py
@@ -2,10 +2,9 @@
 
 import sys
 sys.path.append("../acousticbrainz")
-
+from admin.utils import create_path, remove_old_archives
 import tarfile
 import shutil
-import errno
 import os
 import psycopg2
 import config
@@ -14,43 +13,6 @@ import argparse
 
 DUMP_CHUNK_SIZE = 1000
 
-def create_path(path):
-    """Creates a directory structure if it doesn't exist yet."""
-    try:
-        os.makedirs(path)
-    except OSError as exception:
-        if exception.errno != errno.EEXIST:
-            sys.exit("Failed to create directory structure %s. Error: %s" % (path, exception))
-
-def remove_old_archives(location, pattern, is_dir=False, sort_key=None):
-    """Removes all files or directories that match specified pattern except two last.
-
-    :param location: Location that needs to be cleaned up.
-    :param pattern: Regular expression that will be used to filter entries in the specified location.
-    :param is_dir: True if directories need to be removed, False if files.
-    :param sort_key: See https://docs.python.org/2/howto/sorting.html?highlight=sort#key-functions.
-    """
-    entries = [os.path.join(location, e) for e in os.listdir(location)]
-    pattern = re.compile(pattern)
-    entries = filter(lambda x: pattern.search(x), entries)
-
-    if is_dir:
-        entries = filter(os.path.isdir, entries)
-    else:
-        entries = filter(os.path.isfile, entries)
-
-    if sort_key is None:
-        entries.sort()
-    else:
-        entries.sort(key=sort_key)
-
-    # Leaving only two last entries
-    for entry in entries[:(-2)]:
-        print(' - %s' % entry)
-        if is_dir:
-            shutil.rmtree(entry)
-        else:
-            os.remove(entry)
 
 def dump_lowlevel_json(location=os.path.join(os.getcwd(), 'dump'), rotate=False):
     """
